Refactor network module: consolidate error handling and client functionality by removing the netlink module, introducing a new client structure, and enhancing the wrapper for better abstraction. This improves code organization and facilitates testing.
